int instead of double

+++ variadic_functions/3-print_all.c
@@ -24,7 +24,7 @@ void print_all(const char * const format, ...)
 				printf("%s%c", sep, va_arg(list, int));
 				break;
 				case 'i':
-				printf("%s%f", sep, va_arg(list, double));
+				printf("%s%f", sep, va_arg(list, int));
 				break;
 				case 'f':
 				printf("%s%f", sep, va_arg(list, double));
